feat(test-cases): author the runtime cap in fractional hours
Replace the manifest's `max_runtime_seconds` (integer seconds) with
`max_runtime_hours` (float, default 1), and switch the `--max-runtime`
CLI flag to hours. Every cap is long enough that seconds add no useful
precision, so authoring in hours (e.g. 0.5, 1.5) is clearer.

Parse at the edges, normalize internally: the manifest field and CLI
flag take hours and convert via `runtime_hours_to_seconds`, while the
run pipeline (TestCaseVersion, job API, backend, run records) stays in
seconds — so no contract artifacts change.

Convert all 31 manifests to hours (1800->0.5, 2400->0.75, 3600->1,
5400->1.5, 7200->2, 10800->3, 18000->5) and update the testing manifest
docs, quickstart/setup guides, and authoring skill accordingly.

+++ crates/cli/src/cli.rs
@@ -139,10 +139,11 @@ pub struct RunArgs {
     #[arg(long, value_name = "MODEL")]
     pub model: String,
 
-    /// Maximum harness runtime in seconds before the run is stopped. Overrides
-    /// the test case's `max_runtime_seconds` default; omit to use that default.
-    #[arg(long, value_name = "SECONDS")]
-    pub max_runtime: Option<u64>,
+    /// Maximum harness runtime in hours before the run is stopped. Overrides the
+    /// test case's `max_runtime_hours` default; omit to use that default.
+    /// Fractional hours are allowed (for example `0.5`).
+    #[arg(long, value_name = "HOURS")]
+    pub max_runtime: Option<f64>,
 
     /// Built-in orchestrator that conducts the harness sessions (for example
     /// `one-shot` or `ralph`). Defaults to `one-shot`, a single session. Selection
